# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "bff_message",
    "version": "1.0.0"
  },
  "security": [
    {
      "bearerAuth": []
    },
    {
      "cookieAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "accessToken"
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "enum": [
                  "BAD_REQUEST",
                  "UNAUTHORIZED",
                  "FORBIDDEN",
                  "NOT_FOUND",
                  "CONFLICT",
                  "PAYLOAD_TOO_LARGE",
                  "UNPROCESSABLE_ENTITY",
                  "TOO_MANY_REQUESTS",
                  "INTERNAL_ERROR",
                  "BAD_GATEWAY",
                  "SERVICE_UNAVAILABLE",
                  "GATEWAY_TIMEOUT"
                ]
              },
              "message": {
                "type": "string"
              },
              "details": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "path": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ]
                }
              }
            },
            "required": [
              "code",
              "message",
              "details"
            ]
          }
        },
        "required": [
          "error"
        ]
      },
...
    },
...
  },
  "paths": {
...
    "/contacts": {
      "get": {
        "tags": [
          "Contacts"
        ],
        "summary": "Récupère la liste des contacts",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "description": "Terme de recherche pour filtrer les contacts",
              "example": "Alice"
            },
            "required": false,
            "description": "Terme de recherche pour filtrer les contacts",
            "name": "search",
            "in": "query"
          },
          {
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "description": "Nombre maximum de contacts à retourner",
              "example": 10
            },
            "required": false,
            "description": "Nombre maximum de contacts à retourner",
            "name": "limit",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Liste des contacts",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "contacts": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "anyOf": [
                              {
                                "type": "string"
                              },
                              {
                                "type": "number"
                              }
                            ],
                            "description": "Identifiant unique, peut être une chaîne ou un nombre",
                            "example": "12345"
                          },
                          "name": {
                            "type": "string",
                            "description": "Nom du contact",
                            "example": "Alice Dupont"
                          },
                          "email": {
                            "type": "string",
                            "format": "email",
                            "description": "Adresse email du contact",
                            "example": "alice.dupont@mairie360.fr"
                          },
                          "department": {
                            "type": "string",
                            "description": "Département associé au contact",
                            "example": "Marketing"
                          },
                          "avatarUrl": {
                            "type": "string",
                            "format": "uri",
                            "description": "URL de l’avatar du contact",
                            "example": "https://example.com/avatar.png"
                          },
                          "initials": {
                            "type": "string",
                            "description": "Initiales du contact",
                            "example": "AD"
                          },
                          "presence": {
                            "type": "string",
                            "enum": [
                              "online",
                              "offline",
                              "away"
                            ],
                            "description": "Présence du contact, en ligne, hors ligne ou absent",
                            "example": "online"
                          }
                        },
                        "required": [
                          "id",
                          "name"
                        ],
                        "description": "Représentation d’un contact"
                      },
                      "description": "Liste des contacts"
                    }
                  },
                  "required": [
                    "contacts"
                  ],
                  "description": "Réponse contenant la liste des contacts"
                }
              }
            }
          },
          "400": {
            "description": "Invalid query (details lists the invalid fields)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid session",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "Core API is unavailable or failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}